Support delte by externalIds for Assets, Events, TimeSeries
CDF-13884

+++ src/main/scala/cognite/spark/v1/SdkV1Relation.scala
@@ -158,13 +158,27 @@ abstract class SdkV1Relation[A <: Product, I](config: RelationConfig, shortName:
 
   def deleteWithIgnoreUnknownIds(
       resource: DeleteByIdsWithIgnoreUnknownIds[IO, Long],
-      deletes: Seq[DeleteItem],
-      ignoreUnknownIds: Boolean = true): IO[Unit] = {
-    val ids = deletes.map(_.id)
-    resource
-      .deleteByIds(ids, ignoreUnknownIds)
-      .flatTap(_ => incMetrics(itemsDeleted, ids.length))
-  }
+      ids: Seq[Long],
+      ignoreUnknownIds: Boolean = true): IO[Unit] =
+    if (ids.nonEmpty) {
+      resource
+        .deleteByIds(ids, ignoreUnknownIds)
+        .flatTap(_ => incMetrics(itemsDeleted, ids.length))
+    } else {
+      IO.pure(Unit)
+    }
+
+  def deleteByExternalIdsWithIgnoreUnknown(
+      resource: DeleteByExternalIdsWithIgnoreUnknownIds[IO],
+      externalIds: Seq[String],
+      ignoreUnknownIds: Boolean = true): IO[Unit] =
+    if (externalIds.nonEmpty) {
+      resource
+        .deleteByExternalIds(externalIds, ignoreUnknownIds)
+        .flatTap(_ => incMetrics(itemsDeleted, externalIds.length))
+    } else {
+      IO.pure(Unit)
+    }
 
   def genericUpsert[
       // The Item (read) type
